+++ sdk/src/integrationTest/java/TokenMintIntegrationTest.java
@@ -143,7 +143,7 @@ class TokenMintIntegrationTest {
     void cannotMintTokensWhenSupplyKeyDoesNotSignTransaction() throws Exception {
         var testEnv = new IntegrationTestEnv(1).useThrowawayAccount();
 
-        var key = PrivateKey.generate();
+        var key = PrivateKey.generateED25519();
 
         var response = new AccountCreateTransaction()
             .setKey(key)